base structure

+++ src/utils/data_generator.py
@@ -63,8 +63,8 @@ class DataGenerator(keras.utils.Sequence):
     def __data_generation(self, list_IDs_temp):
         'Generates data containing batch_size samples'  # X : (n_samples, *dim, n_channels)
         # Initialization
-        X = np.zeros((self.batch_size, self.dim), dtype=int)
-        y = np.zeros(self.batch_size, dtype=int)
+        X = np.zeros((self.batch_size, self.dim), dtype=float)
+        y = np.zeros(self.batch_size, dtype=float)
 
         # Generate data
         for i, ID in enumerate(list_IDs_temp):
